// src/app/core/dtos/projects/reports/projectReportOutput.dto.ts
export class projectListRptOutputDto {
  projecT_NUMBER?: string | null = null;
  projecT_NAME?: string | null = null;
  applicatioN_DATE?: Date | null = null;
  arabiC_COUNTRY_NAME?: string | null = null;
  cost?: string | null = null;
  currancY_NAME?: string | null = null;
  local_cost?: string | null = null;
  misC_RECEIPT_AMOUNT?: string | null = null;
  statuS_DESC?: string | null = null;
  projecT_TYPE_DESC?: string | null = null;
  countrY_code?: string | null = null;

  applicatioN_DATEstr?: string | null = null;
  coststr?: string | null = null
  local_coststr?: string | null = null;
  misC_RECEIPT_AMOUNTstr?: string | null = null;
}